requirements.txt
albumentations==0.4.5
black>=19.10b0
fire>=0.2.1
kaggle>=1.5.6
matplotlib>=3.1.2
numpy>=1.14.5
opencv-python>=3.4.2.17
pandas>=0.25.3
pretrainedmodels==0.7.4
pytest>=5.3.2
pytorch_toolbelt==0.3.0
scikit-image>=0.16.2
scikit-learn>=0.19.2
scipy>=1.1.0
segmentation-models-pytorch==0.1.0
tensorflow == 1.15.2
tqdm>=4.25.0 
neptune-client == 0.4.114
numba
psutil>=5.7.0
